Redraw fedora-inverse
[why]
The inverse Fedora logo is drawn by a 'thick' outline. This is valid as
svg, but can not be represented in this way in fonts. It will be filled
and thus generate a logo that is similar to the un-inverted one.

[how]
Remove the outline points (which have the big thickness).
Copy over the fedora.svg and get rid of all points in the middle.
Copy the new thin outline and 'inset' it such, that it in the same
position like the former fat outline.
Join both outlines to form a new filled area outline.
Export as plain SVG.

Fixes #73

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>