Make codepoint allocations persistent and explicit
[why]
When codepoints are changed for the individual logos that should be a
deliberate decision.

[how]
We have now a dedicated file for the codepoint to logo mapping. That
will be used for the font generation: `glyphs.json`.

If new logos show up or exisiting ones are deleted this will change that
dedicated file when `compile.sh` is called and what happened is clearly
visible in the changed git status.

People who add / remove a logo have to run `compile.sh` and afterwards
review the changes in the `glyphs.json` file and include that changes in
their commit.

[note]
This is only slightly different to the formerly used solution: Keep the
`.fontcustom-manifest.json` in git. But that has been removed with
commit f161cf6 for some reason. Maybe because that file is hidden, which
is not so nice.

This PR just preserves the "glyphs" section of the manifest and extracts
it into a dedicated file, that can be added and reviewed.

[note]
Glyph codepoints reverted back to `f161cf6^` i.e. 12f3783 (a.k.a. v0.17)

Fixes #33
Fixes #72

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>