Move Rocky Linux codepoint into empty gap
[why]
For some reason we have two vacant codepoints:
 0xF129 and 0xF12A

Rocky Linux has been appended behind the last previously used codepoint
(on position 0xF12F). That might be suboptimal.

[how]
Reuse vacant slot instead of one in the end...
  0xF12F -> 0xF129

Rocky Linux is only existent in v0.18. With the previous commit we
reverted to the codepoints as per v0.17 and previous. As this would
change the Rocky Linux position in any case we can also put it into a
more optimal position (filling the gap).

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>