Move Alma Linux codepoint into empty gap
[why]
For some reason we have a vacant codepoints:
 0xF12A

Alma Linux has been appended behind the last previously used codepoint
(on position 0xF12E). That might be suboptimal.

[how]
Reuse vacant slot instead of one in the end...
  0xF12E -> 0xF12A

Alma Linux is only existent in v0.18. With commit b69fbf3 we
reverted to the codepoints as per v0.17 and previous. As this would
change the Alma Linux position in any case we can also put it into a
more optimal position (filling the gap).

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>