Sort glyphs.json file by codepoint and not name
[why]
It is easier to see which codepoints are free and how it will all end up
if we sort the table by codepoint instead of by logo name.

[how]
Let jq generate the output in the desired sorting order.
Update glyphs.json file.

Now the glyph.json list will be in ascending codepoint order.
Gaps are more easily visible.
The next codepoint is also easily accessible, just add the new logo in
the bottom.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>