Clean up ferris glyph
[why]
The ferris glyph has multiple outlines on top of each other. That can be
problematic with some rendering engines. It is usually fine for svgs,
but fonts can loose the information what is transparent and what is
black.

[how]
Open in InkScape and just join all paths to ONE path.

The face had another 'background color' (?) with punched out eyes, that
make the joined paths rather complicated (lots of points). To avoid that
the face outline has been cut and re-added below the eyeline.

Export as plain svg. This is not optimized by svgo. I'm not sure this is
wanted here. After ending up in a font it does not make any difference
anyhow.

Fixes: #86

Reported-by: T-727
Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>