CI: Automate release workflow
[why]
Releasing is too hard.

[how]
The maintainer needs to set a new version in the package.json.
Afterwards the release draft creator can be manually triggered on
the Actions tab of the Github repo.

This will draft a release with a new font generated from the current
vectors. And the current commit will be git tagged with the release
version.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>