OpenSCAD: Simplify icon
[why]
The icon has very small details and is hard to see if displayed not big.
This is clearly the multicolor icon that needed color to make sense.
Also need to be bigger.

[how]
Use https://www.svgrepo.com/svg/373950/openscad which is CC0 (they
claim).
Further improve the rendering of the circles and so on, as they have too
many points in the svgrepo version.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>