CI: Enable fetch on commit
[why]
The workflow is more robust when we fetch before we commit new stuff (to
be safe against in-the-meanwhile changes to the repo).

The option to not-fetch has been copied over from another project (Nerd
Fonts) where it made sense because fetching takes a long time due to its
repo size.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>